Even Tree Hackerrank
https://www.hackerrank.com/challenges/even-tree/problem
This is my code to the problem, it is fairly simple. I just calculated each node or subtree's size and those that are even go on into the final answer. Now it was easy also because the type of input is straight froward, and you have to calculate the the size of subtree formed and not subtree remaining. E.g., if u see the sample test code, the edges removed leaves even subtree sizes on both the subtrees formed  whereas my code just checked whether the subtree formed is even or not and still it passes all the test cases. Also - why is syntax coloring not working in github ?